requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.0
Google-GenerativeAI==0.3.1
gunicorn==21.2.0
pytest==7.4.0